+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/dev/ChunkingDemoView.kt
@@ -1,21 +1,18 @@
 package org.wycliffeassociates.otter.jvm.workbookapp.ui.dev
 
-import javafx.beans.property.BooleanProperty
 import javafx.beans.property.IntegerProperty
 import javafx.beans.property.SimpleBooleanProperty
 import javafx.beans.property.SimpleIntegerProperty
-import javafx.scene.control.Button
-import org.kordamp.ikonli.javafx.FontIcon
-import org.kordamp.ikonli.materialdesign.MaterialDesign
 import org.wycliffeassociates.otter.jvm.controls.styles.tryImportStylesheet
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.components.grid.chunkGrid
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.model.ChunkViewData
 import tornadofx.*
 
 class ChunkingDemoView : View() {
-    class ChunkViewData(val number: Int, val completed: BooleanProperty, val selectedChunk: IntegerProperty)
 
     val selectedChunk: IntegerProperty = SimpleIntegerProperty(-1)
 
-    val buttonLabels = listOf(
+    val list = listOf(
         ChunkViewData(1, SimpleBooleanProperty(true), selectedChunk),
         ChunkViewData(2, SimpleBooleanProperty(true), selectedChunk),
         ChunkViewData(3, SimpleBooleanProperty(true), selectedChunk),
@@ -25,32 +22,7 @@ class ChunkingDemoView : View() {
     )
 
     override val root = vbox {
-        gridpane {
-
-            buttonLabels.forEachIndexed { index, chunk ->
-                val btn = Button(chunk.number.toString()).apply {
-                    addClass("btn", "btn--secondary", "btn--borderless", "chunk-item")
-                    graphicProperty().bind(chunk.completed.objectBinding {
-                        this.togglePseudoClass("completed", it == true)
-                        if (it == true) {
-                            FontIcon(MaterialDesign.MDI_CHECK_CIRCLE).apply { addClass("chunk-item__icon") }
-                        } else {
-                            FontIcon(MaterialDesign.MDI_BOOKMARK_OUTLINE).apply { addClass("chunk-item__icon") }
-                        }
-                    })
-
-                    chunk.selectedChunk.onChange {
-                        this.togglePseudoClass("selected", it == chunk.number)
-                    }
-
-                    action {
-                        chunk.selectedChunk.set(chunk.number)
-                    }
-                }
-
-                this.add(btn, index % 3, index / 3)
-            }
-        }
+        chunkGrid(list)
     }
 
     init {
