+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/model/ChunkViewData.kt
@@ -1,0 +1,6 @@
+package org.wycliffeassociates.otter.jvm.workbookapp.ui.model
+
+import javafx.beans.property.BooleanProperty
+import javafx.beans.property.IntegerProperty
+
+class ChunkViewData(val number: Int, val completed: BooleanProperty, val selectedChunk: IntegerProperty)
